+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Unreleased
+
+## Added
+
+## Fixed
+
+## Changed
+
 # 0.43.186-alpha (2025-04-24 / 6b1ecae)
 
 ## Added