+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 
 ## Changed
 
+- Make sure the clojure process can communicate over `stdio`.
+
 # 0.42.182-alpha (2025-03-01 / 50cc172)
 
 ## Fixed